DABOM-323 refactor: 파일 구조 이름
