+++ apps/service/src/api/auth/schema.ts
